requirements.txt
# Core Dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Data Visualization
plotly==5.18.0
altair==5.2.0

# API and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0

# Database
sqlite3